+++ inst/builtin-templates/bslib-bare/modules/output_widgets/server.R
@@ -1,0 +1,10 @@
+library(shiny)
+library(shidashi)
+
+server <- function(input, output, session, ...) {
+
+  server_base_plot(input, output, session, ...)
+  server_ggplot_brush(input, output, session, ...)
+  server_threebrain(input, output, session, ...)
+
+}